Create README.md
This is a multi-class classification project from a kaggle dataset which consists of various types of features related to a particular song on spotify (audio features and track metadata). This project aims to classify a song into 3 categories- Low popular, Moderately popular and Highly Popular. The code has been mentioned in the next file accordingly.